% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/Kirby21_data.R
\name{Kirby 21 Data}
\alias{Kirby 21 Data}
\title{Kirby 21 Visit 1 data}
\source{
\url{https://www.nitrc.org/frs/?group_id=313}
}
\description{
A dataset containing 3 images (FLAIR, MPRAGE, T2w) for 21 subjects
from Kennedy Krieger
}

